Move content-api to modules
