+++ controllers/home/index.js
@@ -1,7 +1,7 @@
 'use strict';
 const path = require('path');
 const express = require('express');
-const contentApi = require('../../services/content-api');
+const contentApi = require('../../modules/content-api');
 
 const router = express.Router();
